issue #392: 戦略B demand paging の code review 指摘 (15件) を修正
PR #396 (Phase 4 4a〜4f + alloc_huge) に対する max-effort code review の findings を
まとめて修正。全て EMULIN_NATIVE_PF opt-in 経路に限定され、default off (eager) は
byte-identical (native-oracle 代表6件 + native-pf-oracle で確認)。

correctness:
- #1 ≥2GB munmap の int 切り詰め: MemoryBackend.free / Memory.free / NativeMemoryBackend.free を
  long size 化、sys_munmap / amd64_mremap の (int) cast 撤去。128GB cage の munmap が no-op 化せず
  removeReserveRange + 物理回収が走る。
  ★#1 を直すと V8 cage と VA が重なる guest stack を cage-trim munmap が巻き込んで unmap し stack を
  破壊する潜在問題が露見 (pre-fix は munmap no-op で隠れていた)。→ stack 帯 [stackBottom-8MB,
  stackBottom) を free() は unmap せず faultIn が demand grow する保護を追加 (Linux stack auto-grow 相当)。
- #2 per-vCPU TSS 64-slot 衝突: vcpuId は KVM id として monotonic 必須なので、TSS band slot を別の
  recyclable な tssSlot に分離 (teardown で free-list へ返却、>64 同時で clean に throw)。生涯 64 thread
  超で tssBase が KSTACK_BASE と衝突する triple fault を解消。
- #3 runtime MAP_FIXED が予約帯 clobber: NativeCpuBackend が [GDT,KSTACK band] を setReservedBand で
  登録、anonMmap の fixed 経路が踏んだら relocate (TSS/IDT 破壊 DoS を防止)。
- #4 realloc in-place grow が reserve-only 隣接を見落とす: virt2phys(present) に加え overlapsReserve で
  reserve 衝突も relocate (重複 entry / 物理共有 corruption を防止)。
- #5 KvmVcpu.getCr2 の per-#PF arena.allocate leak: sregsBuf を ctor 確保し再利用 (getCr2/getCpl/
  exitToRing3/configure* 共通、WHP と同方針)。
- #6 faultIn の scan<64 cap: maxReserveLen-bounded 下方走査 (containingReserve) で深い nesting の
  包含 region 取りこぼし (spurious SIGSEGV) を解消。
- #7 removeReserveRange/overlapsReserve の単一 floorEntry: 重なる lower entry を全て (bounded) 処理。
- #8 alloc_huge の overlap 無検査: addr!=0 hint が reserve と衝突したら kernel-chooses に fallback。
- #9 mmapTop underflow: bumpDown で MMAP_FLOOR/underflow guard (heap/text 帯侵入を NativeOom 化)。
- #13 free() が ≥2GB を全ページ probe: absentStride で PML4/PDPT/PD level の absent stride を skip。

efficiency / conventions:
- #15 EMULIN_TRACE_MMAP を static final TRACE_MMAP に (CLAUDE.md「debug env flag は static final」)。
- #14 (readGprs before #PF 判定): KVM に単一 RIP read が無く naive 分割は syscall path を悪化させるため
  コメントで意図的保留 (WHP 限定の微小最適化)。
- #10/#11/#C5 (prot 非追跡の residual): mmapRegions は範囲のみ追跡し prot を持たない件を faultIn に明記
  (prot-aware VMA 追跡は別 refactor)。#16 (alloc_huge ⊃ anonMmap dup): per-page ループが 128GB で
  hang するため delegate 不可と明記 (妥当な重複)。

test: sys_pf_huge64 のコメントを #1 後の munmap 実行を反映して更新。
検証: claude --version が default 512MB pool + NATIVE_PF で完走、native-pf-oracle 2/2、
native(eager) 代表6件 (hello/mmap/static/fork/fork_exec/dyn) == software。

+++ src/main/java/emulin/Memory.java
@@ -636,7 +636,7 @@ public class Memory extends Elf implements MemoryBackend
   //   worker 並走時に free の alloclist.remove + alloclistGen++ が alloc/resolve と
   //   race しうる。兄弟 mutator と同じ monitor で囲み get→remove→gen++→buf=null を
   //   atomic 化する (intrinsic lock は reentrant なので alloclist 保持経路から呼ばれても安全)。
-  public int free( long address, int size ) {
+  public int free( long address, long size ) {   // issue #392 review #1: size を long 化 (≥2GB 切り詰め防止)
     synchronized( alloclist ) {
       AllocInfo allocinfo = alloclist.get( address );
       if( allocinfo == null || allocinfo.size != size ) return -1;
